Fix generation of random points in unit sphere
The calculation of random points in the unit sphere
uses a normal distribution rather than a uniform distribution.
When using a uniform distribution, the points were generated near
the poles of the spheres. By switching to a normal distribution,
the points are now generated uniformly throughout the sphere.

The implementation change requires a new thread_rng generator to be
initialized for each pixel and passed to the functions that use
the random_point_in_sphere function.